+++ packages/signals/lib/signals.dart
@@ -1,33 +1,3 @@
 library signals;
 
-export 'src/connect.dart' show connect, Connect;
-export 'src/extensions/future.dart';
-export 'src/extensions/iterable.dart';
-export 'src/extensions/list.dart';
-export 'src/extensions/map.dart';
-export 'src/extensions/set.dart';
-export 'src/extensions/stream.dart';
-export 'src/core/signals.dart'
-    show
-        signal,
-        computed,
-        effect,
-        batch,
-        untracked,
-        disableSignalsDevTools,
-        EffectCleanup,
-        Computed,
-        Signal,
-        ValueSignal,
-        ReadonlySignal;
-export 'src/utils/timer.dart'
-    show timerSignal, TimerSignal, TimerSignalDurationUtils;
-export 'src/async/async_state.dart';
-export 'src/async/async_signal.dart' show AsyncSignal;
-export 'src/async/stream_signal.dart' show streamSignal, StreamSignal;
-export 'src/async/future_signal.dart' show futureSignal, FutureSignal;
-export 'src/value/iterable_signal.dart' show iterableSignal, IterableSignal;
-export 'src/value/list_signal.dart' show listSignal, ListSignal;
-export 'src/value/map_signal.dart' show mapSignal, MapSignal;
-export 'src/value/set_signal.dart' show setSignal, SetSignal;
-export 'src/core/observer.dart';
+export 'package:signals_core/signals_core.dart';