+++ packages/signals/lib/signals_flutter.dart
@@ -1,9 +1,3 @@
 library signals_flutter;
 
-export 'signals.dart';
-export 'src/extensions/signal.dart';
-export 'src/extensions/value_listenable.dart';
-export 'src/extensions/value_notifier.dart';
-export 'src/watch/widget.dart';
-export 'src/watch/extension.dart'
-    hide getSignalListenSubscriberCount, getSignalWatchSubscriberCount;
+export 'package:signals_flutter/signals_flutter.dart';