+++ packages/signals_core/test/value_signals_test.dart
@@ -1,5 +1,5 @@
-import 'package:signals/signals.dart';
-import 'package:flutter_test/flutter_test.dart';
+import 'package:signals_core/signals_core.dart';
+import 'package:test/test.dart';
 
 void main() {
   group('value signals test', () {