+++ packages/signals_flutter/lib/signals_flutter.dart
@@ -1,0 +1,9 @@
+library signals_flutter;
+
+export 'package:signals_core/signals_core.dart';
+export 'src/extensions/signal.dart';
+export 'src/extensions/value_listenable.dart';
+export 'src/extensions/value_notifier.dart';
+export 'src/watch/widget.dart';
+export 'src/watch/extension.dart'
+    hide getSignalListenSubscriberCount, getSignalWatchSubscriberCount;