+++ packages/signals_flutter/lib/src/extensions/value_listenable.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/foundation.dart';
-
-import '../core/signals.dart';
+import 'package:signals_core/signals_core.dart';
 
 /// Extension on [ValueListenable] to provide helpful methods for signals
 extension SignalValueListenableUtils<T> on ValueListenable<T> {